+++ website/static/playground/worker.js
@@ -56,6 +56,8 @@ async function instantiate() {
         host_dom_storage_get: () => 0,
         host_dom_storage_set: () => {},
         host_dom_storage_remove: () => {},
+        host_dom_state_get: () => 0,
+        host_dom_state_set: () => {},
         host_dom_worker_spawn: () => 0n,
         host_dom_worker_send: () => {},
         host_dom_worker_on: () => {},